reduce to functions
