+++ tests/rosrust_msg_test.rs
@@ -1,9 +1,7 @@
 #[cfg(feature = "rosrust_msg")]
 #[test]
 fn convertxyz_rosrust_msg() {
-    use ros_pointcloud2::{
-        pcl_utils::PointXYZ, reader::ReaderXYZ, writer::WriterXYZ, PointCloud2Msg,
-    };
+    use ros_pointcloud2::{pcl_utils::PointXYZ, PointCloud2Msg};
 
     let cloud = vec![
         PointXYZ {
@@ -23,10 +21,10 @@ fn convertxyz_rosrust_msg() {
         },
     ];
     let copy = cloud.clone();
-    let internal_cloud: PointCloud2Msg = WriterXYZ::from(cloud).try_into().unwrap();
+    let internal_cloud = PointCloud2Msg::try_from_iterable(cloud).unwrap();
     let rosrust_msg_cloud: rosrust_msg::sensor_msgs::PointCloud2 = internal_cloud.into();
     let convert_back_internal: PointCloud2Msg = rosrust_msg_cloud.into();
-    let to_convert = ReaderXYZ::try_from(convert_back_internal).unwrap();
+    let to_convert = convert_back_internal.try_into_iter().unwrap();
     let back_to_type = to_convert.collect::<Vec<PointXYZ>>();
     assert_eq!(copy, back_to_type);
 }
